% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{addsToCart}
\alias{addsToCart}
\title{'Adds-To-Cart'}
\format{
A data frame with 12 rows and 3 columns:
\describe{
\item{dim_year}{Calendar Year}
\item{dim_month}{Calendar Month}
\item{addsToCart}{Total Quantity of Items Added to Cart - aggregated by Calendar Year & Month}
}
}
\description{
'Adds-To-Cart'
}
