% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTable.R
\name{calcDescRank}
\alias{calcDescRank}
\title{Descending Rank of Column Values}
\usage{
calcDescRank(data, x)
}
\arguments{
\item{data}{data.frame}

\item{x}{character}
}
\value{
data.frame
}
\description{
Descending Rank of Column Values
}
\examples{
\dontrun{
  calcDescRank(
    data = sessionCounts, 
    x = c('sessions', 'transactions', 'QTY')
  )
}
}
