% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.colname}
\alias{is.colname}
\title{Check if vector contains valid column names}
\usage{
is.colname(data, x, scalar = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{x}{character}

\item{scalar}{logical}
}
\value{
logical
}
\description{
Check if vector contains valid column names
}
\examples{
\dontrun{
 is.colname(data.frame(a = 1, b = 2), x = c('a', 'b'))
 is.colname(data.frame(z = 1), x = c('z'), scalar = TRUE)
}
}
