% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTable.R
\name{pctSumDesc}
\alias{pctSumDesc}
\title{Percentage of Total Sum (Ordered by Rank of Descending Values)}
\usage{
pctSumDesc(data, x)
}
\arguments{
\item{data}{data.frame}

\item{x}{character - must be length 1}
}
\value{
data.frame
}
\description{
Percentage of Total Sum (Ordered by Rank of Descending Values)
}
\examples{
\dontrun{
  pctSumDesc(data = sessionCounts, x = 'sessions')
  pctSumDesc(data = sessionCounts, x = 'transactions')
  pctSumDesc(data = sessionCounts, x = 'QTY')
}
}
