% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sessionCounts}
\alias{sessionCounts}
\title{'Session Counts'}
\format{
A data frame with 7,734 rows and 6 columns:
\describe{
\item{dim_browser}{Name of Web Browser}
\item{dim_deviceCategory}{Category of Device}
\item{dim_date}{Calendar Date for Browsing Activity}
\item{sessions}{Total # of Web Browsing Sessions - aggregated by Calendar Date, Browser and Device Category}
\item{transactions}{Total # of Transactions - aggregated by Calendar Date, Browser and Device Category }
\item{QTY}{Total Quantity of Purchased Items - aggregated by Calendar Date, Browser and Device Category}
}
}
\description{
'Session Counts'
}
