% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTable.R
\name{toParetoTable}
\alias{toParetoTable}
\title{Summarize Data in Pareto Table}
\usage{
toParetoTable(data = NULL, dims = NULL, vals = NULL, desc_by = NULL, agg = sum)
}
\arguments{
\item{data}{data.frame}

\item{dims}{character}

\item{vals}{character}

\item{desc_by}{character - must be length 1}

\item{agg}{function}
}
\value{
data.frame
}
\description{
Summarize Data in Pareto Table
}
\examples{
\dontrun{
  toParetoTable(
    data = sessionCounts, 
    dims = 'dim_browser', 
    vals = c('sessions', 'transactions', 'QTY'), 
    desc_by = 'sessions'
  )
}
}
