% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretoTable.R
\name{cumPctSumDesc}
\alias{cumPctSumDesc}
\title{Cumulative Percentage of Total Sum}
\usage{
cumPctSumDesc(data, x)
}
\arguments{
\item{data}{data.frame}

\item{x}{character - must be length 1}
}
\value{
data.frame
}
\description{
Cumulative Percentage of Total Sum
}
\examples{
\dontrun{
  cumPctSumDesc(data = sessionCounts, x = 'sessions')
  cumPctSumDesc(data = sessionCounts, x = 'transactions')
  cumPctSumDesc(data = sessionCounts, x = 'QTY')
}
}
